requirements.txt
Click==7.0
Flask==1.0.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pyaml==19.4.1
PyYAML==5.1.1
Werkzeug==0.15.4
